4.21-4.22

+++ osa4/blogilista/controllers/blogs.js
@@ -1,4 +1,3 @@
-const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
@@ -21,11 +20,7 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
-    const decodedToken = jwt.verify(request.token, process.env.SECRET)
-    if (!decodedToken.id) {
-        return response.status(401).json({ error: 'token invalid' })
-    }
-    const user = await User.findById(decodedToken.id)
+    const user = request.user
     const blog = new Blog({
         title: body.title,
         author: body.author,
@@ -40,9 +35,18 @@ blogsRouter.post('/', async (request, response) => {
     response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response, next) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+blogsRouter.delete('/:id', async (request, response) => {
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+    const userId = blog.user.toString()
+    console.log("user.id: ", user.id)
+    console.log("userId: ", userId)
+    if(user.id === userId) {
+        await Blog.findByIdAndDelete(request.params.id)
+        response.status(204).end()
+    } else {
+        return response.status(401).json({ error: 'token invalid' })
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
